move achievement vdom to separate component
