+++ services/app/apps/codebattle/assets/css/style.scss
@@ -377,11 +377,6 @@ a {
   overflow: hidden;
 }
 
-.cb-profile {
-  width: 230px;
-  height: auto;
-}
-
 .cb-username {
   font-size: 27px;
 }
